+++ crate-common/src/main/java/me/oriley/crate/SvgAsset.java
@@ -1,0 +1,11 @@
+package me.oriley.crate;
+
+@SuppressWarnings("unused")
+public class SvgAsset extends Asset {
+
+    // TODO: Info retrieval
+
+    SvgAsset(String path, String name) {
+        super(path, name);
+    }
+}