+++ crate-runtime/src/main/java/me/oriley/crate/Crate.java
@@ -2,12 +2,12 @@ package me.oriley.crate;
 
 import android.content.Context;
 import android.content.res.AssetManager;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.Typeface;
+import android.graphics.*;
+import android.graphics.drawable.PictureDrawable;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.Log;
+import me.oriley.crate.CrateSvg.SvgParseException;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -19,6 +19,8 @@ import java.util.HashMap;
 @SuppressWarnings("unused")
 public final class Crate {
 
+    private static final String TAG = Crate.class.getSimpleName();
+
     @NonNull
     private final AssetManager mAssetManager;
 
@@ -29,6 +31,12 @@ public final class Crate {
 
     @NonNull
     private final HashMap<String, Bitmap> mBitmapCache = new HashMap<>();
+
+    @NonNull
+    private final HashMap<String, PictureDrawable> mSvgCache = new HashMap<>();
+
+    @NonNull
+    private final CrateSvg.Parser mSvgParser = CrateSvg.getParser();
 
     @NonNull
     public final CrateDictionary mDictionary;
@@ -61,10 +69,10 @@ public final class Crate {
         if (mBitmapCache.containsKey(key)) {
             Bitmap bitmap = mBitmapCache.get(key);
             if (bitmap != null && !bitmap.isRecycled()) {
-                if (DEBUG) Log.d("Crate", "Using cached bitmap for key: " + key);
+                if (DEBUG) Log.d(TAG, "Using cached bitmap for key: " + key);
                 return bitmap;
             } else {
-                if (DEBUG) Log.d("Crate", "Ejecting recycled bitmap for key: " + key);
+                if (DEBUG) Log.d(TAG, "Ejecting recycled bitmap for key: " + key);
                 mBitmapCache.remove(key);
             }
         }
@@ -78,11 +86,11 @@ public final class Crate {
                 stream.close();
             }
         } catch (IOException e) {
-            Log.e("Crate", "Failed to load bitmap for key: " + key, e);
+            Log.e(TAG, "Failed to load bitmap for key: " + key, e);
             e.printStackTrace();
         } finally {
             if (bitmap != null) {
-                if (DEBUG) Log.d("Crate", "Bitmap loaded for key: " + key);
+                if (DEBUG) Log.d(TAG, "Bitmap loaded for key: " + key);
                 mBitmapCache.put(key, bitmap);
             }
         }
@@ -93,27 +101,81 @@ public final class Crate {
     public Typeface getTypeface(@NonNull FontAsset fontAsset) {
         String key = fontAsset.mPath;
         if (mTypefaceCache.containsKey(key)) {
-            if (DEBUG) Log.d("Crate", "Using cached typeface for key: " + key);
+            if (DEBUG) Log.d(TAG, "Using cached typeface for key: " + key);
             return mTypefaceCache.get(key);
         }
         Typeface typeface = null;
         try {
             typeface = Typeface.createFromAsset(mAssetManager, key);
         } catch (RuntimeException e) {
-            Log.e("Crate", "Failed to load typeface for key: " + key, e);
+            Log.e(TAG, "Failed to load typeface for key: " + key, e);
             e.printStackTrace();
         } finally {
             if (typeface != null) {
-                if (DEBUG) Log.d("Crate", "Typeface loaded for key: " + key);
+                if (DEBUG) Log.d(TAG, "Typeface loaded for key: " + key);
                 mTypefaceCache.put(key, typeface);
             }
         }
         return typeface;
     }
 
+    @Nullable
+    public Bitmap getSvgBitmap(@NonNull SvgAsset svgAsset) {
+        return getSvgBitmap(svgAsset, 1f);
+    }
+
+    @Nullable
+    public Bitmap getSvgBitmap(@NonNull SvgAsset svgAsset, float scale) {
+        String key = svgAsset.mPath;
+
+        PictureDrawable pictureDrawable = getSvgDrawable(svgAsset);
+        if (pictureDrawable == null) {
+            if (DEBUG) Log.d(TAG, "PictureDrawable is null for key: " + key);
+            return null;
+        }
+
+        int desiredWidth = (int) (pictureDrawable.getIntrinsicWidth() * scale);
+        int desiredHeight = (int) (pictureDrawable.getIntrinsicHeight() * scale);
+        Bitmap bitmap = Bitmap.createBitmap(desiredWidth, desiredHeight, Bitmap.Config.ARGB_8888);
+        Canvas canvas = new Canvas(bitmap);
+        canvas.drawPicture(pictureDrawable.getPicture(), new Rect(0, 0, desiredWidth, desiredHeight));
+
+        return bitmap;
+    }
+
+    @Nullable
+    public PictureDrawable getSvgDrawable(@NonNull SvgAsset svgAsset) {
+        String key = svgAsset.getPath();
+        if (mSvgCache.containsKey(key)) {
+            if (DEBUG) Log.d(TAG, "Using cached PictureDrawable for key: " + key);
+            return mSvgCache.get(key);
+        }
+
+        PictureDrawable drawable = null;
+        try {
+            InputStream stream = open(svgAsset);
+            //noinspection TryFinallyCanBeTryWithResources
+            try {
+                drawable = mSvgParser.parseSvg(stream);
+            } finally {
+                stream.close();
+            }
+        } catch (IOException | SvgParseException e) {
+            Log.e(TAG, "Failed to load SVG for key: " + key, e);
+            e.printStackTrace();
+        } finally {
+            if (drawable != null) {
+                if (DEBUG) Log.d(TAG, "SVG loaded for key: " + key);
+                mSvgCache.put(key, drawable);
+            }
+        }
+        return drawable;
+    }
+
     public void clear() {
         clearBitmapCache();
         clearTypefaceCache();
+        clearSvgCache();
     }
 
     public void clearBitmapCache() {
@@ -121,6 +183,10 @@ public final class Crate {
             bitmap.recycle();
         }
         mBitmapCache.clear();
+    }
+
+    public void clearSvgCache() {
+        mSvgCache.clear();
     }
 
     public void clearTypefaceCache() {